requirements.txt
flask
flask-sqlalchemy
flask-migrate
flask-cors
flask-jwt-extended
flask-mail
flask-limiter
requests
celery
redis
python-dotenv
psycopg2-binary
gunicorn
requests
werkzeug
